pyproject.toml
[project]
name = "garage-discordbot"
version = "0.1.0"
description = "Discord bot for Jim's Garage"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorlog>=6.9.0",
    "discord>=2.3.2",
    "dotenv>=0.9.9",
    "fastapi>=0.115.11",
    "pyyaml>=6.0.2",
    "uvicorn>=0.34.0",
]

[dependency-groups]
dev = [
    "pre-commit>=4.1.0",
    "pyre-check>=0.9.23",
    "pyright>=1.1.394",
    "ruff>=0.9.7",
    "shellcheck-py>=0.10.0.1",
    "yamllint>=1.35.1",
]

# Note - this is not yet consumed by Pyre.  A recent commit enables it, but there has not been a fresh build of Pyre yet:
# See https://github.com/facebook/pyre-check/issues/695
[tool.pyre]
ignore_all_errors = [".venv"]
site_package_search_strategy="pep561"
source_directories = ["."]
strict = true

[tool.ruff]
indent-width = 4
line-length = 88
target-version = "py313"

[tool.ruff.lint]
select = ["ALL"]
ignore = [
    "COM812",
    "D"
]
# COM812 - ruff recommends disabling it for compatibilitiy (and will warn you about it every time)
# D - docustring checking

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
indent-style = "space"
quote-style = "double"
